Implemented Hack to Show Checked Off Services
Summary:

Currently editing the initial set for TransactionUpdate,
which renders RepairsForm to app/edit.html.

The problem is that by default, TransactionUpdate renders
the checkboxes in edit.html by setting their values, but
leaving their "checked" attribute to the default (default=checked)

Solving this issue by editing the initial set is debatedly okay,
if it can be done by iterating over the desired fields in RepairsForm.
This could potentially be all fields besides price and service
description.

However, the desired behavior is actually to DISPLAY exactly the checkboxes
that were originally checked when creating the transaction. So we should
only display a checkbox for frame if e.g. need_handlebars is true in the
model (this was set when creating transaction). We should then check it
off or not based on done_handlebars (this can be set and reset when
editing a transaction).