Core Transaction Functionality works, with one bug
Bug: All Transaction checkbox fields are displayed,
even though we only want to display those fields
that were checked when the Transaction was originally
created.

Reason: Unchecked checkboxes are not posted, which is the
same as checkboxes which are hidden in the template and
therefore not posted. Both are getting handled the same way-
by setting the field to "IN_PROGRESS", which is only correct
in the former case.

Possible Fixes:
-Figure out how to not update ModelForm fields that are None
in the database.
-Figure out how to get the primary key in the form, so that
we can query the database for each unchecked field and update
its value to its current value (set it to itself).