Implemented Rental and Refurbished Bike Features
--Create and View Rental and Refurbished Bikes--

Rental and Refurbished functionalities are the same. Rental is described below:

Link on index page takes user to /rental page

The /rental displays all bikes in the RentalBike table.
Clicking on one of the bikes' vin number takes the user to
a detail page /rental_detail, which takes as a url argument
the rental bike primary key id.

The /rental_detail displays all transactions in the transaction_set
of transactions which have foreign keys to this rental bike.

The /rental page also allows users to add a rental bike.

This takes users to the /new_rental page, where the template renders
the ModelForm associated with RentalBike. Right now this is just composed
of a VIN number. Upon submitting, the new rental bike is added to the table.

--When Creating a Transaction--

There is now a field for Rental or Refurbished bike VIN number.

If the user specifies a VIN number for a valid rental/refurbished bike,
the Transaction sets its foreign key to the associated RentalBike.
If the user specifies an invalid VIN number, no error is thrown at the moment, but the Transaction keeps its foreign key null.