Fixed task creation/display for the Complete Task List
There were two problems:

(1) rendering task names containing white spaces as id's in the templates was causing those div's to throw errors and not display at all. Solution was to apply a filter to replace " " with "_". When displaying task fields, the reverse filter was used to replace "_" with " ".

(2) In info_dict, the category keys mapped to the field name values. These values instead needed to match the category keys from category_dict. In edit.html, the info_dict category values are checked against the category_dict keys.