Fixed dropdown data rendering by move to Models
Moving choices to models still saves the same value, but supplying
a choices attribute to the model field enables Django to create
a get_<model_field>_display method to render the human-readable
choice in the template.

Also excluded the transaction field from the PartCategoryForm,
since this is automatically set to the transaction being created.