Added was_used field to PartCategory
Implementation

In edit.html, used trick to submit checkbox when it is false as well as when it is true. Did this by having a hidden second input with the same name as the checkbox. The name attribute is the key to the value posted to the server in the form.

The rest happens in edit.js. If the checkbox is not checked, do nothing and the hidden second input will overwrite the first. If the checkbox is checked, disable the second input and the checked checkbox will be included in the form, like normal.

Added the was_used field to PartCategory in models.py, and saved the MultiDict was_used values to their PartCategory instances in views.py.