Completed transition to MenuItem from hardcoded dicts
__SUMMARY__

MAJOR CHANGES:
MenuItem's are the choices the customer has for servicing his bike.
The fields for MenuItems are Name, Category, and Price.

There is no need for Task's to contain name, category, or price information.
They are just order instances of a MenuItem which are associated with a Transaction.
I changed their fields to reflect this.
The fields are now Completed, Transaction, and MenuItem

This change enables Brian or another admin to change the Menu, via the /admin page.
The hard-coded dicts are now completely removed from the code (though info_dict is
present in /app/fixtures/seed_menu.json, where it is used to seed the database.

MINOR CHANGES:
I removed TasksForm completely, and made RepairsForm a ModelForm only with the
Cost and Service_Description fields of the Transaction model. This removes the
need to filter out the task fields of RepairsForm when rendering to /new. Now
all two of the RepairsForm fields are rendered with no filtering.

FURTHER POSSIBLE REFACTORING TO DO:
On the create_transaction page (/new), a dictionary mapping categories to lists
of MenuItems is passed to the template.
But on the edit page, a list of categories is passed in, along with a list of
MenuItems. Then the MenuItems are sorted client-side with some jQuery. There is
no reason for this. It is probably better design to do this sorting server-side
just like on the /new page, i.e. by passing in a dictionary mapping categories
to lists of MenuItems.