Modified /assign_parts to disable editing
Added a DisabledPartCategory form. Assigned category forms are created in
pairs when the user clicks "Add Category". The displayed form has all fields
disabled (DisabledPartCategory form), and a hidden form is the one actually
submitted. Using a disabled form was necessary, because the readonly attribute
does not work for select fields in html.