Styled balance page
Used bootstrap btn-group div to justify anchor tag link with submit button.
That refers to the "add" and "export" buttons.

Added css @media selector to hide .non-mobile class dom elements. On the balance
page, I hid the Customer column for screen widths < 480px. Iphone5 is 320px in
width.

I removed padding from the add button, so that if someone were to increase the
.button-pic height from 25px to up to about 40px, the add button would still
line up with the Export button.